returns a sorted array of article objects - tests reconsolidated into two test suites
