+++ app.js
@@ -1,5 +1,5 @@
 const { getTopics } = require("./controllers/topics.controller");
-const { getArticleById, patchArticleVotesById } = require("./controllers/articles.controller");
+const { getArticleById, patchArticleVotesById, getArticles } = require("./controllers/articles.controller");
 const { getUsers } = require("./controllers/users.controller");
 
 const express = require('express');
@@ -10,6 +10,7 @@ app.use(express.json())
 app.get('/api/topics', getTopics);
 
 //articles
+app.get('/api/articles',getArticles);
 app.get('/api/articles/:article_id',getArticleById);
 app.patch('/api/articles/:article_id',patchArticleVotesById);
