+++ package.json
@@ -38,7 +38,7 @@
   },
   "jest": {
     "setupFilesAfterEnv": [
-      "jest-extended/all"
+      "jest-extended/all","jest-sorted"
     ]
   }
 }
